+++ test/plausible_web/controllers/api/stats_controller/conversions_test.exs
@@ -12,8 +12,8 @@ defmodule PlausibleWeb.Api.StatsController.ConversionsTest do
       conn = get(conn, "/api/stats/#{site.domain}/conversions?period=day&date=2019-01-01")
 
       assert json_response(conn, 200) == [
-               %{"name" => "Signup", "count" => 3, "total_count" => 3},
-               %{"name" => "Visit /register", "count" => 2, "total_count" => 2}
+               %{"name" => "Signup", "count" => 3, "total_count" => 3, "meta_keys" => ["variant"]},
+               %{"name" => "Visit /register", "count" => 2, "total_count" => 2, "meta_keys" => nil}
              ]
     end
   end
@@ -34,8 +34,29 @@ defmodule PlausibleWeb.Api.StatsController.ConversionsTest do
         )
 
       assert json_response(conn, 200) == [
-               %{"name" => "Signup", "count" => 3, "total_count" => 3}
+               %{"name" => "Signup", "count" => 3, "total_count" => 3, "meta_keys" => ["variant"]}
              ]
+    end
+  end
+
+  describe "GET /api/stats/:domain/meta-breakdown/:key" do
+    setup [:create_user, :log_in, :create_site]
+
+    test "returns metadata breakdown for goal", %{conn: conn, site: site} do
+      insert(:goal, %{domain: site.domain, event_name: "Signup"})
+      filters = Jason.encode!(%{goal: "Signup"})
+      meta_key = "variant"
+
+      conn =
+        get(
+          conn,
+          "/api/stats/#{site.domain}/meta-breakdown/#{meta_key}?period=day&date=2019-01-01&filters=#{filters}"
+        )
+
+      assert json_response(conn, 200) == [
+        %{"count" => 2, "name" => "B", "total_count" => 2},
+        %{"count" => 1, "name" => "A", "total_count" => 1}
+      ]
     end
   end
 end